实现新数据入库分类

+++ src/main/java/casia/isiteam/militaryqa/common/Constant.java
@@ -44,7 +44,7 @@ public class Constant {
     public static final String Nature_Keywords = "keywords";
 
     /** 二级类别与c_id映射，用于新数据分类 */
-    public static Map<String, Integer> smallCategoriesMapping = new HashMap<>();
+    public static Map<String, Long> smallCategoriesMapping = new HashMap<>();
 
 
     static {
